% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{compliance}
\alias{compliance}
\title{Calculate standard compliance for each area}
\usage{
compliance(
  envdata,
  exclude_stores = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  standard = "BS 4971",
  min_temp = FALSE,
  max_temp = FALSE,
  min_RH = FALSE,
  max_RH = FALSE
)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger.}

\item{start_date}{A string to use as a minimum date.}

\item{end_date}{A string to use as a maximum date.}

\item{standard}{Standard ("BS 4971", "PAS 198", "Icon", or "Bizot").}

\item{min_temp}{Minimum temperature, overwrites standard if provided.}

\item{max_temp}{Maximum temperature, overwrites standard if provided.}

\item{min_RH}{Minimum RH, overwrites standard if provided.}

\item{max_RH}{Maximum RH, overwrites standard if provided.}
}
\value{
rated A dataframe with overall standard compliance, percentage of time above and below, and mean temp and RH.
}
\description{
Calculate standard compliance for each area
}
\examples{
compliance(envdata, standard = "Icon")
}
