% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_compliance}
\alias{graph_compliance}
\title{Graph standard compliance}
\usage{
graph_compliance(
  envdata,
  type = "o",
  standard = "BS 4971",
  exclude_stores = FALSE,
  graph_title = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  min_temp = FALSE,
  max_temp = FALSE,
  min_RH = FALSE,
  max_RH = FALSE,
  col_low = "#336699",
  col_good = "#669933",
  col_high = "#993322",
  grad_bad = "#990000",
  grad_mid = "#CC6600",
  grad_good = "#006600"
)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger}

\item{type}{'o' for overall graph, 't' for temperature, or 'r' for RH.}

\item{standard}{A string for graph titles and to set minimum and maximum values if recognised.}

\item{exclude_stores}{A string or vector to exclude a group of stores.}

\item{graph_title}{Title of graph, else 'All stores at site' or first location if \code{store} specified.}

\item{start_date}{A string to use as minimum date.}

\item{end_date}{A string to use as maximum date.}

\item{min_temp}{Minimum temperature guideline.}

\item{max_temp}{Maximum temperature guideline.}

\item{min_RH}{Minimum RH guideline.}

\item{max_RH}{Maximum RH guideline.}

\item{col_low}{Colour of low bar for temp or RH graph, as named HTML colour or hex.}

\item{col_good}{Colour of good bar for temp or RH graph, as named HTML colour or hex.}

\item{col_high}{Colour of high bar for temp or RH graph, as named HTML colour or hex.}

\item{grad_bad}{Colour of low end of gradient for overall graph, as named HTML colour or hex.}

\item{grad_mid}{Colour of midpoint of gradient for overall graph, as named HTML colour or hex.}

\item{grad_good}{Colour of high end of gradient for overall graph, as named HTML colour or hex.}
}
\value{
ggplot2 Object containing bar graph of overall or temp/RH rating for each space in dataframe
}
\description{
Graph standard compliance
}
\examples{
graph_compliance(envdata, 't', standard = 'Icon')
}
