% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_light}
\alias{graph_light}
\title{Graph light data}
\usage{
graph_light(
  envdata,
  store = FALSE,
  exclude_stores = FALSE,
  graph_title = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  breaks = "2 months",
  date_format = "\%m/\%y",
  max_lux = 50,
  max_UV = 0,
  col_lux = "darkgreen",
  col_UV = "darkorange"
)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger.}

\item{store}{A string matching a single store or a group of stores with a matching pattern.}

\item{exclude_stores}{A string or vector to exclude a group of stores.}

\item{graph_title}{Title of graph, else 'All stores at site' or first location if \code{store} specified.}

\item{start_date}{A string to use as minimum date.}

\item{end_date}{A string to use as maximum date.}

\item{breaks}{Breaks on x axis, like '2 weeks'.}

\item{date_format}{Format of dates on x axis, like '\%d/\%m'.}

\item{max_lux}{Value of maximum lux guideline.}

\item{max_UV}{Value of maximum UV guideline}

\item{col_lux}{Color for visible light line, as named HTML color or hex.}

\item{col_UV}{Color for UV line, as named HTML color or hex.}
}
\value{
ggplot2 Object containing visible light and UV graph
}
\description{
Graph light data
}
\examples{
\dontrun{
 graph_light(envdata_exhib, graph_title = 'Exhibition cases')
}

}
