% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_light_summary}
\alias{graph_light_summary}
\title{Graph light data summary}
\usage{
graph_light_summary(
  envdata,
  graph_title = FALSE,
  store = FALSE,
  exclude_stores = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  type = "monthly",
  percentile = FALSE,
  breaks = "2 months",
  date_format = "\%m/\%Y",
  max_lux = 50,
  max_UV = 0,
  col_lux = "darkgreen",
  col_UV = "darkorange"
)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger.}

\item{graph_title}{Title of graph, else 'All stores at site' or first location if \code{store} specified.}

\item{store}{A string matching a single store or a group of stores with a matching pattern.}

\item{exclude_stores}{A string or vector to exclude a group of stores.}

\item{start_date}{A string to use as minimum date.}

\item{end_date}{A string to use as maximum date.}

\item{type}{'annual', 'monthly', or 'daily'}

\item{breaks}{Breaks on x axis, like '2 weeks'.}

\item{date_format}{Format of dates on x axis, like '\%d/\%m' or '12/24'.}

\item{max_lux}{Value of maximum lux guideline.}

\item{max_UV}{Value of maximum UV guideline}

\item{col_lux}{Colour for visible light line and ribbon, as named HTML colour or hex.}

\item{col_UV}{Colour for UV line and ribbon, as named HTML colour or hex.}
}
\value{
ggplot2 Object containing visible light and UV graph with mean and min/max ribbons
}
\description{
Graph light data summary
}
\examples{
graph_light_summary(envdata_exhib, graph_title = 'Exhibition cases', type = 'daily')
}
