% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_move}
\alias{graph_move}
\title{Splice temp and RH graph on specified date}
\usage{
graph_move(
  envdata1,
  envdata2,
  store1,
  store2,
  move_date,
  graph_title = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  date_format = "\%m/\%Y",
  breaks = "2 months",
  standard = "BS 4971",
  min_temp = FALSE,
  max_temp = FALSE,
  min_RH = FALSE,
  max_RH = FALSE,
  max_axis_temp = 40,
  max_axis_RH = 100,
  col_temp = "red",
  col_RH = "blue",
  col_line = "darkgrey"
)
}
\arguments{
\item{envdata1}{Dataframe returned from parse_datalogger for left of graph}

\item{envdata2}{Dataframe returned from parse_datalogger for right of graph}

\item{store1}{String to search location in envdata1}

\item{store2}{String to search location in envdata2}

\item{move_date}{Date to splice data}

\item{graph_title}{Title of graph, otherwise 'Store 1 to Store 2'.}

\item{start_date}{A string to use as minimum date.}

\item{end_date}{A string to use as maximum date.}

\item{date_format}{Format of dates on x axis, like '\%d/\%m'.}

\item{breaks}{Breaks on x axis, like '2 weeks'.}

\item{standard}{Standard to use for temp/RH guidelines.}

\item{min_temp}{Minimum temperature guideline.}

\item{max_temp}{Maximum temperature guideline.}

\item{min_RH}{Minimum RH guideline.}

\item{max_RH}{Maximum RH guideline.}

\item{max_axis_temp}{Maximum of temperature axis.}

\item{max_axis_RH}{Maximum of RH axis.}

\item{col_temp}{Colour for temperature line, as named HTML colour or hex.}

\item{col_RH}{Colour for RH line, as named HTML colour or hex.}

\item{col_line}{Colour for vertical line at splice date}
}
\value{
ggplot2 Object containing temp and RH graph of two sets of data spliced on \code{move_date}.
}
\description{
Splice temp and RH graph on specified date
}
\examples{
graph_move(envdata_old, envdata_new, store1 = 'Archive 1', store2 = 'Archive A', move_date = '2025-01-01')
}
