% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_store}
\alias{graph_store}
\title{Graph temperature and RH for single or multiple stores}
\usage{
graph_store(
  envdata,
  store = FALSE,
  exclude_stores = FALSE,
  graph_title = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  date_format = "\%m/\%Y",
  breaks = "2 months",
  standard = "BS 4971",
  min_temp = FALSE,
  max_temp = FALSE,
  min_RH = FALSE,
  max_RH = FALSE,
  max_axis_temp = 40,
  max_axis_RH = 100,
  col_temp = "red",
  col_RH = "blue"
)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger.}

\item{store}{A string matching a single store or a group of stores with a matching pattern.}

\item{exclude_stores}{A string or vector to exclude a group of stores.}

\item{graph_title}{Title of graph, else 'All stores at site' or first location if \code{store} specified.}

\item{start_date}{A string to use as minimum date.}

\item{end_date}{A string to use as maximum date.}

\item{date_format}{Format of dates on x axis, like '\%d/\%m'.}

\item{breaks}{Breaks on x axis, like '2 weeks'.}

\item{standard}{Standard to use for temp/RH guidelines.}

\item{min_temp}{Minimum temperature guideline.}

\item{max_temp}{Maximum temperature guideline.}

\item{min_RH}{Minimum RH guideline.}

\item{max_RH}{Maximum RH guideline.}

\item{max_axis_temp}{Maximum of temperature axis.}

\item{max_axis_RH}{Maximum of RH axis.}

\item{col_temp}{Colour for temperature line, as named HTML colour or hex.}

\item{col_RH}{Colour for RH line, as named HTML colour or hex.}
}
\value{
ggplot2 Object containing temperature and RH graph.
}
\description{
Graph temperature and RH for single or multiple stores
}
\examples{
graph_store(envdata_exhib, graph_title = 'Exhibition cases')
}
