% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{remove_faulty}
\alias{remove_faulty}
\title{Remove potentially faulty readings outside range}
\usage{
remove_faulty(envdata, min_temp = -25, max_temp = 35, min_RH = 10, max_RH = 90)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger}

\item{min_temp}{A number for the minimum realistic temperature}

\item{max_temp}{A number for the maximum realistic temperature}

\item{min_RH}{A number for the minimum realistic RH}

\item{max_RH}{A number for the maximum realistic RH}
}
\value{
subset A dataframe with rows out of range removed
}
\description{
Remove potentially faulty readings outside range
}
\examples{
remove_faulty(envdata, max_RH = 70)
}
