% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{set_minmax}
\alias{set_minmax}
\title{Set minimum and maximum temp with standard or custom values}
\usage{
set_minmax(
  standard = "BS 4971",
  min_temp = FALSE,
  max_temp = FALSE,
  min_RH = FALSE,
  max_RH = FALSE
)
}
\arguments{
\item{standard}{Standard ("BS 4971", "PAS 198", "Icon", or "Bizot")}

\item{min_temp}{Minimum temperature, overwrites standard if provided}

\item{max_temp}{Maximum temperature, overwrites standard if provided}

\item{min_RH}{Minimum RH, overwrites standard if provided}

\item{max_RH}{Maximum RH, overwrites standard if provided}
}
\value{
minmax A vector of minimum and maximum values for temp and RH
}
\description{
Set minimum and maximum temp with standard or custom values
}
\examples{
set_minmax(standard = "PAS 198", max_temp = 25)
}
