% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_site}
\alias{summarize_site}
\title{Summarize environmental data with maximum, minimum, range, median, and average temperature and RH. Optional 1-99th percentile.}
\usage{
summarize_site(
  envdata,
  exclude_stores = FALSE,
  start_date = FALSE,
  end_date = FALSE,
  type = "monthly",
  percentile = FALSE
)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger}

\item{exclude_stores}{A string or list of stores to exclude from analysis}

\item{start_date}{A string to use as minimum date}

\item{end_date}{A string to use as maximum date}

\item{type}{'annual', 'monthly', or 'daily'}
}
\value{
site_summary A dataframe of summary statistics
}
\description{
Summarize environmental data with maximum, minimum, range, median, and average temperature and RH. Optional 1-99th percentile.
}
\examples{
dplyr::summarize_site(envdata, start_date = '2024-01-01', type = 'annual')
}
