% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{light_dose}
\alias{light_dose}
\title{Calculate light dose and compare to blue wool standards}
\usage{
light_dose(envdata, start_date = FALSE, end_date = FALSE, obs_hour = 3)
}
\arguments{
\item{envdata}{A dataframe returned from parse_datalogger}

\item{start_date}{A string to use as a minimum date}

\item{end_date}{A string to use as a maximum date}

\item{obs_hour}{Number of observations logged per hour}
}
\value{
dose A dataframe of the cumulative lux and UV hours with percent to just noticeable fade for BW standards
}
\description{
Calculate light dose and compare to blue wool standards
}
\examples{
light_dose(envdata_light, obs_hour = 4)
}
