% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_brand}
\alias{parse_brand}
\title{parse_brand}
\usage{
parse_brand(
  filepath,
  brand = FALSE,
  site = "",
  model = "",
  location = "",
  old_data = NULL
)
}
\arguments{
\item{filepath}{Path to a directory of logfiles of the same brand}

\item{brand}{The logger brand in lowercase}

\item{site}{The site name used for each row as a string for any brand other than Meaco}

\item{model}{Model as a string for brands that do not include the model in the file. Not required for TinyTag}

\item{location}{Location of the logger if not recoverable from the file}

\item{old_data}{Dataframe containing previously parsed data to append}
}
\description{
A central function to route to the appropriate specific logger extraction function
}
\examples{
\dontrun{
 parse_datalogger("/data/tinytag/2025-01-01_to_2025-02-01.csv",
"Anonymous Library", "tinytag")
}
}
