% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{subset_readings}
\alias{subset_readings}
\title{Subset readings}
\usage{
subset_readings(
  envdata,
  store = FALSE,
  exclude_stores = FALSE,
  start_date = FALSE,
  end_date = FALSE
)
}
\arguments{
\item{envdata}{Dataframe returned from parse_brand}

\item{store}{Store name or pattern to match in the location}

\item{exclude_stores}{Store name or pattern in the location to filter out}

\item{start_date}{Date in "YYYY-mm-dd" format}

\item{end_date}{Date in "YYYY-mm-dd" format}
}
\value{
subset, a filtered dataframe
}
\description{
Subset readings
}
\examples{
subset_readings(envdata, store = "C", start_date = "2024-01-01", end_date = "2024-12-31")
}
